vixen: fold the HTTP origin into vixen-primitives, delete the rest
`vix-fetch` read as "the crate that implements fetch". It did not: the fetch
PRIMITIVE has always been in vixen-primitives beside decode, tree-read and
registry-url. What lived in vix-fetch was one 30-line `OriginAdapter` — two
guard clauses, `ureq::get`, read the body — and 255 lines that nothing used.

So the transport moves in as an ordinary module, `ureq` becomes an ordinary
dependency, and the crate is gone. What the boundary was buying was keeping a
TLS stack out of the primitive layer; that is a real trade, and Amos took it
deliberately rather than hiding it behind a feature flag.

Deleted rather than migrated:

- `FetchBackend`/`FetchOutput`/`HttpArchiveFetchBackend` — no consumers anywhere,
  and the backend unpacked archives itself, which
  `machine.primitive.fetch-returns-a-blob` forbids ("an archive is a file;
  unpacking is a separate demand"). Its own test file went with it: three tests
  for a dead contract, one of which fetched itoa from static.crates.io, so the
  suite no longer needs the network to pass.
- the flat `Map<Path, Blob>` tree — the recursive model with directories,
  symlinks and executable bits replaced it, and this was the last hold-out.

The 657-line origin integration test moves to `vixen-runtime/tests/`, following
the precedent set when the dev-dep cycle was broken: tests that drive programs
through `vixen_runtime::ratchet` are system tests and live with the harness.
Putting them in vixen-primitives would have recreated exactly the cycle that
commit removed, which `cargo tree --edges all` gates in CI.

430 tests pass, down from 433 by precisely the three dead-contract tests. The
`vix-core-bare` gate still holds: nothing about this touches the language crate.
