--- vix-fetch/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "vix-fetch"
-version = "0.0.0"
-edition = "2024"
-publish = false
-description = "Real fetch backend for vix archives."
-
-[package.metadata]
-
-[package.metadata."docs.rs"]
-rustdoc-args = ["--html-in-header", "arborium-header.html"]
-
-[dependencies]
-vix = { path = "../vix-core", package = "vix-core" }
-flate2 = "1.1"
-tar = "0.4"
-ureq = "3.3"
-sha2.workspace = true
-hex.workspace = true
-
-[dev-dependencies]
-tempfile.workspace = true
-# The fetch/observe integration test drives programs through the ratchet harness,
-# which lives in the vixen runtime (it installs the builtin primitives).
-vixen-runtime = { path = "../vixen-runtime" }
